ref(fzf-lua): simplify implementation
The previous implementation works best when streaming something. Here,
contents are already in memory. Since that is already out the window,
I'm more concerned with maintainability.

Rather than rebase, I want to keep the old implementation in history.
This way it can be reverted to in case it is needed again.