feat: fzf-lua integration (#47)
* feat(config): add settings for fzf-lua

I tried to add a minimal number of settings by default to reduce maintainer burden.
However, the settings I did expose hopefully provide the options that most users want.

* feat(picker): fzf-lua support

You can read more about how I did this here:
https://github.com/ibhagwan/fzf-lua/wiki/Advanced

The hardest part was the previewer, and threading through the snippetsByDisplayName
to the previewer. But in the grand scheme of things... not all that complicated.

* feat(picker): detect if fzf-lua is installed, and use it if so

I followed how snacks and telescope were implemented here.

* docs: fzf-lua config settings

* docs: update install instructions

This shows users that fzf-lua is an option when installing the plugin.

I tried to stay consistent with what was there before... but it seemed like there was a mix
of tabs and spaces here? Was difficult to get it right.

* ref(fzf-lua): simplify implementation

The previous implementation works best when streaming something. Here,
contents are already in memory. Since that is already out the window,
I'm more concerned with maintainability.

Rather than rebase, I want to keep the old implementation in history.
This way it can be reverted to in case it is needed again.

* fix(fzf-lua): use type hints that the CI can resolve